flask-api/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
flask-mysqldb==0.2.0
marshmallow-sqlalchemy==0.26.0
connexion==2.7.0
sshtunnel==0.4.0
requests==2.25.1